--HG--
branch : Windows

+++ TameParse/Lr/lalr_builder.cpp
@@ -65,7 +65,7 @@ typedef map<item_container, lalr_state_container> state_for_item;
 void lalr_builder::create_closure(closure_set& target, const lalr_state& state, const grammar* gram) {
     queue<lr1_item_container> waiting;
     
-    for (size_t itemId = 0; itemId < state.count_items(); itemId++) {
+    for (int itemId = 0; itemId < state.count_items(); itemId++) {
         // Mark this item as waiting
         lr0_item_container  lr0 = state[itemId];
         lr1_item_container  lr1(new lr1_item(state[itemId], state.lookahead_for(itemId)), true);
@@ -135,7 +135,7 @@ void lalr_builder::complete_parser() {
             int         offset  = (*item)->offset();
             
             // Items at the end of a rule don't produce any transitions
-            if (offset == rule.items().size()) continue;
+            if (offset == (int) rule.items().size()) continue;
             
             // Get the item that the 'dot' is before
             const item_container& dottedItem = rule.items()[offset];
@@ -235,7 +235,7 @@ void lalr_builder::complete_lookaheads() {
         const lalr_machine::transition_set& transitions = m_Machine.transitions_for_state(stateId);
         
         // Iterate through the items in this state
-        for (size_t itemId = 0; itemId < thisState->count_items(); itemId++) {
+        for (int itemId = 0; itemId < thisState->count_items(); itemId++) {
             // Get the item
             const lalr_state::container& thisItem = (*thisState)[itemId];
